finished all changes, untested
